+++ lib/src/annotations.dart
@@ -1,6 +1,36 @@
-class JsonKey {
+class DeserializationInfo {
+  final String key;
+  final Map<String, dynamic> map;
+  final Object? current;
+
+  const DeserializationInfo({
+    required this.key,
+    required this.map,
+    this.current,
+  });
+}
+
+class SerializationInfo {
+  final String key;
+  final Object? data;
+
+  const SerializationInfo({
+    required this.key,
+    this.data,
+  });
+}
+
+class JsonKey<T> {
   final String name;
-  const JsonKey(this.name);
+
+  final T Function(DeserializationInfo info)? fromJson;
+  final Object? Function(SerializationInfo info)? toJson;
+
+  const JsonKey(
+    this.name, {
+    this.fromJson,
+    this.toJson,
+  });
 }
 
 class JsonIgnore {